refactor(spec): split schema.ts into names, declarations, catalog, facade
Separate TypeBox name registries, the ordered $defs bag, pipeline catalogs,
and Cyclic/Module assembly so domain edits do not require reading the full
schema module. Public ./schema.js exports and schema/v0.json are unchanged.
